admin page broken

+++ backend/src/main/java/com/techelevator/model/BeerDAO.java
@@ -1,6 +1,6 @@
 package com.techelevator.model;
 
-
+import java.math.BigDecimal;
 import java.util.List;
 
 //TODO Brewery model will need to use beerId to display beer list
@@ -9,7 +9,7 @@ public interface BeerDAO {
 	
 	public List<Beer> getAllBeer();
 	
-	public Beer getBeerById(Long id);
+	public Beer getBeerById(int id);
 
 	public Beer getBeerByName(String name);
 	
@@ -27,5 +27,6 @@ public interface BeerDAO {
 	//public boolean searchForBeerByName(String name);
 	
 	//public void updateBeerAvailability(Long beerId, boolean isActive);
+	
 	
 }
